Change: team_id & user_id to unsignedBigInteger in team_users Migration

+++ Modules/Teams/database/migrations/2024_08_03_145931_create_team_users_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('team_users', function (Blueprint $table) {
             $table->id();
-            $table->integer('team_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('team_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
 
             //? Foreign Key
